3DRRadio support for MultiPoint firmware
The new MultiPoint firmware, "MP SiK 2.5", responds with a prefix of the
Node ID : "[0] " for node 0. These changes effectively ignore that.
Tested with "SiK 1.6" to ensure backwards compatibility, only with
HM-TRP hardware.

The MP firmware seems to have issues with remote commands, but this can
be used to set each one individually.  Next will be the RSSI screen,
which may require a firmware tweak I already did.

There are also tweaks to the UI, including a "History" log to show the
commands and responses. Logging was cleaned up a little.

NOTE: This was done with VS2015, which altered some of the .resx files
significantly, with the UI tweaks. The actual protocol support is only
in 3DRadio.cs